refactor(vyuh_cli): file name clean up
