+++ vyuh_cli/lib/src/commands/commands.dart
@@ -1,2 +1,2 @@
-export 'create/create.dart';
-export 'update_command.dart';
+export 'create/create_command.dart';
+export 'update/update_command.dart';
