+++ vyuh_cli/lib/src/commands/create/commands/feature/feature.dart
@@ -1,5 +1,5 @@
 import 'package:vyuh_cli/src/commands/create/commands/feature/feature_app_subcommand.dart';
-import 'package:vyuh_cli/src/commands/create/templates/templates.dart';
+import 'package:vyuh_cli/src/commands/create/templates/index.dart';
 
 class CreateFeature extends FeatureSubCommand {
   CreateFeature({
