+++ vyuh_cli/lib/src/commands/create/commands/flutter_app/index.dart
