+++ vyuh_cli/lib/src/commands/create/templates/index.dart
@@ -1,0 +1,3 @@
+export 'template.dart';
+export 'vyuh_feature/index.dart';
+export 'vyuh_flutter_app/index.dart';
