+++ vyuh_cli/lib/src/commands/create/templates/template.dart
@@ -1,28 +1,18 @@
 import 'package:mason/mason.dart';
 import 'package:universal_io/io.dart';
 
-/// {@template template}
-/// Dart class that represents a Vyuh supported template.
-/// Each template consists of a [MasonBundle], name,
-/// and help text describing the template.
-/// {@endtemplate}
 abstract class Template {
-  /// {@macro template}
   const Template({
     required this.name,
     required this.bundle,
     required this.help,
   });
 
-  /// The name associated with this template.
   final String name;
 
-  /// The [MasonBundle] used to generate this template.
   final MasonBundle bundle;
 
-  /// The help text shown in the usage information for the CLI.
   final String help;
 
-  /// Callback invoked after template generation has completed.
   Future<void> onGenerateComplete(Logger logger, Directory outputDir);
 }
