--- vyuh_cli/lib/src/commands/create/templates/templates.dart
@@ -1,3 +1,0 @@
-export 'template.dart';
-export 'vyuh_feature/vyuh_feature.dart';
-export 'vyuh_flutter_app/vyuh_flutter_app.dart';
