+++ vyuh_cli/lib/src/commands/create/templates/vyuh_flutter_app/index.dart
