+++ vyuh_cli/lib/src/commands/update/update_command.dart
